Add files via upload
This file is shows the map drawn using power BI tool. The excel file contains the stations located in San Joaquin valley with other parameters like elevation of each station. The map drawn shows the elevation value at each station. The legend is displayed next to the map.
Power query was used to import and load and close the excel file. Then power map was used to graphically represent the elevation at each station. 2 layers are added to the map. The first layer just plots the stations on the map. The second layer is used to color the map depending on the elevation at each station.
The files attached are as follows:
1. Chill Portion Summary: This excel file is the input data that is used for graphical representation.
1. Power BI heat map on elevation:  This is the final outcome after running the power map tool in excel.